+++ src/vegapunk/simulators/fetorch/element/type/tetra4.py
@@ -104,7 +104,7 @@ class FETetra4(ElementType):
                           (1.0, 0.0, 0.0),
                           (0.0, 1.0, 0.0),
                           (0.0, 0.0, 1.0)],
-                         dtype=torch.float, device=self._device)
+                         device=self._device)
         # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
         # Store nodes local coordinates
         self._nodes_local_coords = nodes_local_coords
